Testing enums

+++ tests/Assert/AssertTests.cpp
@@ -7,38 +7,38 @@ namespace nyan
 {
     TEST(Assert, Test) {
         using namespace assert;
-        constexpr auto testAssert = Assert<AssertionLevel::Enabled, AssertionExitMode::Abort, AssertionLogMode::Disabled>{};
+        constexpr auto testAssert = Assert< static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Abort), static_cast<size_t>(AssertionLogMode::Disabled)>{};
         EXPECT_EXIT({
             testAssert(false);
             }, testing::ExitedWithCode(3), "");
 
-        constexpr auto testAssertMsgOnly = Assert<AssertionLevel::Enabled, AssertionExitMode::Abort, AssertionLogMode::MessageOnly>{};
+        constexpr auto testAssertMsgOnly = Assert< static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Abort), static_cast<size_t>(AssertionLogMode::MessageOnly )> {};
         EXPECT_EXIT({
             testAssertMsgOnly(false, "Hello");
             }, testing::ExitedWithCode(3), "Hello");
 
-        constexpr auto testSourceLoc = Assert<AssertionLevel::Enabled, AssertionExitMode::Abort, AssertionLogMode::SourceLocation>{};
+        constexpr auto testSourceLoc = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Abort), static_cast<size_t>(AssertionLogMode::SourceLocation )> {};
         EXPECT_EXIT({
             testSourceLoc(false);
             }, testing::ExitedWithCode(3), "nyan::Assert");
 
-        constexpr auto testAssertStackTrace = Assert<AssertionLevel::Enabled, AssertionExitMode::Abort, AssertionLogMode::StackTrace>{};
+        constexpr auto testAssertStackTrace = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Abort), static_cast<size_t>(AssertionLogMode::StackTrace )> {};
         EXPECT_EXIT({
             testAssertStackTrace(false);
             }, testing::ExitedWithCode(3), "Test::Run");
 
 
-        constexpr auto testAssertExit = Assert<AssertionLevel::Enabled, AssertionExitMode::Exit, AssertionLogMode::Disabled>{};
+        constexpr auto testAssertExit = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Exit), static_cast<size_t>(AssertionLogMode::Disabled )> {};
         EXPECT_EXIT({
             testAssertExit(false);
             }, testing::ExitedWithCode(1), "");
 
-        constexpr auto testAssertQuickExit = Assert<AssertionLevel::Enabled, AssertionExitMode::QuickExit, AssertionLogMode::Disabled>{};
+        constexpr auto testAssertQuickExit = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::QuickExit), static_cast<size_t>(AssertionLogMode::Disabled )> {};
         EXPECT_EXIT({
             testAssertQuickExit(false);
             }, testing::ExitedWithCode(1), "");
 
-        constexpr auto testAssertNoExit = Assert<AssertionLevel::Enabled, AssertionExitMode::Disabled, AssertionLogMode::Disabled>{};
+        constexpr auto testAssertNoExit = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Disabled), static_cast<size_t>(AssertionLogMode::Disabled )> {};
         EXPECT_NO_FATAL_FAILURE({
             testAssertNoExit(false);
             });
@@ -46,7 +46,7 @@ namespace nyan
     TEST(Assert, Lazy) {
 
         using namespace assert;
-        constexpr auto testAssert = Assert<AssertionLevel::Enabled, AssertionExitMode::Disabled, AssertionLogMode::Disabled>{};
+        constexpr auto testAssert = Assert < static_cast<size_t>(AssertionLevel::Enabled), static_cast<size_t>(AssertionExitMode::Disabled), static_cast<size_t>(AssertionLogMode::Disabled )> {};
 
         int flag = 0;
 
@@ -58,7 +58,7 @@ namespace nyan
 
         flag = 0;
 
-        constexpr auto testAssertDisabled = Assert<AssertionLevel::Disabled, AssertionExitMode::Disabled, AssertionLogMode::Disabled>{};
+        constexpr auto testAssertDisabled = Assert < static_cast<size_t>(AssertionLevel::Disabled), static_cast<size_t>(AssertionExitMode::Disabled), static_cast<size_t>(AssertionLogMode::Disabled)>{};
 
         EXPECT_NO_FATAL_FAILURE({
             testAssertDisabled([&]() {flag++; return false; });
